--- code/system.ui/react.common/src/common/const.COLORS.mts
@@ -1,17 +1,0 @@
-export const COLORS = {
-  BLACK: '#000',
-  WHITE: '#fff',
-
-  DARK: '#293042', // Inky blue/black.
-  CYAN: '#00C2FF',
-  MAGENTA: '#FE0064',
-
-  BLUE: '#4D7EF7',
-  GREEN: '#00BB47',
-  YELLOW: '#FCC400',
-
-  LIME: '#A6E130',
-  PURPLE: '#A11D8E',
-
-  RED: '#E21B22',
-} as const;
