+++ code/system.ui/react.common/src/common/index.ts
@@ -1,6 +1,6 @@
 import type * as t from './types.mjs';
 
-export { t };
+export type { t };
 export * from '../index.pkg.mjs';
 export * from './libs.mjs';
 export * from './const.mjs';
